⚡ Adding node input to support chat request
